// frontend/src/openapi/model/trabajadorDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TrabajadorDto { 
    id?: number;
    nombre?: string;
    apellidos?: string;
    dni?: string;
    naf?: string;
    genero?: TrabajadorDto.GeneroEnum;
    fechaNacimiento?: string;
    fechaAntiguedad?: string;
    categoria?: string;
    iban?: string;
    telefono?: number;
    email?: string;
    password?: string;
    rol?: TrabajadorDto.RolEnum;
    activo?: boolean;
    empresa_id?: number;
}
export namespace TrabajadorDto {
    export type GeneroEnum = 'FEMENINO' | 'MASCULINO' | 'NO_DETERMINADO';
    export const GeneroEnum = {
        Femenino: 'FEMENINO' as GeneroEnum,
        Masculino: 'MASCULINO' as GeneroEnum,
        NoDeterminado: 'NO_DETERMINADO' as GeneroEnum
    };
    export type RolEnum = 'ADMIN' | 'EMPRESA' | 'TRABAJADOR';
    export const RolEnum = {
        Admin: 'ADMIN' as RolEnum,
        Empresa: 'EMPRESA' as RolEnum,
        Trabajador: 'TRABAJADOR' as RolEnum
    };
}